Refactor BudgetLine
