+++ app/models/concerns/gobierto_budgets/budget_line_elasticsearch_helpers.rb
@@ -1,0 +1,409 @@
+module GobiertoBudgets
+  module BudgetLineElasticsearchHelpers
+    extend ActiveSupport::Concern
+
+    class_methods do
+
+      def where(conditions)
+        validate_conditions(conditions)
+        @conditions = conditions
+        self
+      end
+
+      def first
+        terms = [
+          {term: { kind: @conditions[:kind] }},
+          {term: { year: @conditions[:year] }},
+          {term: { code: @conditions[:code] }},
+          {missing: { field: 'functional_code'}},
+          {missing: { field: 'custom_code'}},
+          {term: { ine_code: @conditions[:place].id }}
+        ]
+
+        query = {
+          sort: [
+            { @sort_attribute => { order: @sort_order } }
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          size: 10_000
+        }
+
+        area     = BudgetArea.klass_for(@conditions[:area_name])
+
+        response = SearchEngine.client.search(
+          index: SearchEngineConfiguration::BudgetLine.index_forecast,
+          type: area.area_name,
+          body: query
+        )
+
+        hits = response['hits']['hits']
+
+        raise BudgetLine::RecordNotFound if hits.empty?
+
+        BudgetLinePresenter.new(hits.first['_source'].merge(
+          site: @conditions[:site],
+          kind: @conditions[:kind],
+          area: area
+        ))
+      end
+
+      def functional_codes_for_economic_budget_line(conditions)
+        terms = [
+          {term: { kind: conditions[:kind] }},
+          {term: { year: conditions[:year] }},
+          {term: { code: conditions[:functional_code] }},
+          {exists: { field: 'functional_code'}},
+          {term: { ine_code: conditions[:place].id }}
+        ]
+
+        query = {
+          sort: [
+            { @sort_attribute => { order: @sort_order } }
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          aggs: {
+            total_budget: { sum: { field: 'amount' } },
+            total_budget_per_inhabitant: { sum: { field: 'amount_per_inhabitant' } },
+          },
+          size: 10_000
+        }
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast,
+                                                               type: EconomicArea.area_name, body: query
+
+        response['hits']['hits'].map{ |h| h['_source'] }.map do |row|
+          next if row['functional_code'].length != 1
+          area = FunctionalArea
+          row['code'] = row['functional_code']
+
+          BudgetLinePresenter.new(row.merge(kind: EXPENSE, area: area, site: @conditions[:site]))
+        end.compact.sort{|b,a| a.amount <=> b.amount }
+      end
+
+      def all
+        terms = [
+          {term: { kind: @conditions[:kind] }},
+          {term: { ine_code: @conditions[:place].id }}
+        ]
+
+        terms.push({term: { year: @conditions[:year] }}) if @conditions[:year]
+        terms.push({term: { code: @conditions[:code] }}) if @conditions[:code]
+        terms.push({term: { level: @conditions[:level] }}) if @conditions[:level]
+        terms.push({term: { parent_code: @conditions[:parent_code] }}) if @conditions[:parent_code]
+        if @conditions[:functional_code]
+          if @conditions[:area_name] == FunctionalArea.area_name
+            @conditions[:area_name] = EconomicArea.area_name
+            terms.push({term: { functional_code: @conditions[:functional_code] }})
+          else
+            @conditions[:area_name] = FunctionalArea.area_name
+            return functional_codes_for_economic_budget_line(@conditions)
+          end
+        elsif @conditions[:custom_code]
+          if @conditions[:area_name] == CustomArea.area_name
+            @conditions[:area_name] = EconomicArea.area_name
+            terms.push({term: { custom_code: @conditions[:custom_code] }})
+          # else
+          #   @conditions[:area_name] = CustomArea.area_name
+          #   return functional_codes_for_economic_budget_line(@conditions)
+          end
+        else
+          terms.push({missing: { field: 'functional_code'}})
+          terms.push({missing: { field: 'custom_code'}})
+        end
+
+        query = {
+          sort: [
+            { @sort_attribute => { order: @sort_order } }
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          aggs: {
+            total_budget: { sum: { field: 'amount' } },
+            total_budget_per_inhabitant: { sum: { field: 'amount_per_inhabitant' } },
+          },
+          size: 10_000
+        }
+
+        area = BudgetArea.klass_for(@conditions[:area_name])
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast,
+                                                               type: area.area_name, body: query
+
+        response['hits']['hits'].map{ |h| h['_source'] }.map do |row|
+          BudgetLinePresenter.new(row.merge(
+            site: @conditions[:site],
+            kind: @conditions[:kind],
+            area: area,
+            total: response['aggregations']['total_budget']['value'],
+            total_budget_per_inhabitant: response['aggregations']['total_budget_per_inhabitant']['value']
+          ))
+        end
+      end
+
+      def search(options)
+
+        terms = [{term: { kind: options[:kind] }},
+                {term: { year: options[:year] }}]
+
+        terms << {term: { ine_code: options[:ine_code] }} if options[:ine_code].present?
+        terms << {term: { parent_code: options[:parent_code] }} if options[:parent_code].present?
+        terms << {term: { level: options[:level] }} if options[:level].present?
+        terms << {term: { code: options[:code] }} if options[:code].present?
+        if options[:range_hash].present?
+          options[:range_hash].each_key do |range_key|
+            terms << {range: { range_key => options[:range_hash][range_key] }}
+          end
+        end
+
+        query = {
+          sort: [
+            { code: { order: 'asc' } }
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          aggs: {
+            total_budget: { sum: { field: 'amount' } },
+            total_budget_per_inhabitant: { sum: { field: 'amount_per_inhabitant' } },
+          },
+          size: 10_000
+        }
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast, type: (options[:type] || EconomicArea.area_name), body: query
+
+        return {
+          'hits' => response['hits']['hits'].map{ |h| h['_source'] },
+          'aggregations' => response['aggregations']
+        }
+      end
+
+      def for_ranking(options)
+        response = budget_line_query(options)
+        results = response['hits']['hits'].map{|h| h['_source']}
+        total_elements = response['hits']['total']
+
+        return results, total_elements
+      end
+
+      def place_position_in_ranking(options)
+        id = %w{ine_code year code kind}.map {|f| options[f.to_sym]}.join('/')
+
+        response = budget_line_query(options.merge(to_rank: true))
+        buckets = response['hits']['hits'].map{|h| h['_id']}
+        position = buckets.index(id) ? buckets.index(id) + 1 : 0;
+        return position
+      end
+
+      def budget_line_query(options)
+
+        terms = [
+          {term: { year: options[:year] }},
+          {term: { kind: options[:kind] }},
+          {term: { code: options[:code] }}
+        ]
+
+        if options[:filters].present?
+          population_filter =  options[:filters][:population]
+          total_filter = options[:filters][:total]
+          per_inhabitant_filter = options[:filters][:per_inhabitant]
+          aarr_filter = options[:filters][:aarr]
+        end
+
+        if (population_filter && (population_filter[:from].to_i > Population::FILTER_MIN || population_filter[:to].to_i < Population::FILTER_MAX))
+          reduced_filter = {population: population_filter}
+          reduced_filter.merge!(aarr: aarr_filter) if aarr_filter
+          results,total_elements = Population.for_ranking(options[:year], 0, nil, reduced_filter)
+          ine_codes = results.map{|p| p['ine_code']}
+          terms << [{terms: { ine_code: ine_codes }}] if ine_codes.any?
+        end
+
+        if (total_filter && (total_filter[:from].to_i > BudgetTotal::TOTAL_FILTER_MIN || total_filter[:to].to_i < BudgetTotal::TOTAL_FILTER_MAX))
+          terms << {range: { amount: { gte: total_filter[:from].to_i, lte: total_filter[:to].to_i} }}
+        end
+
+        if (per_inhabitant_filter && (per_inhabitant_filter[:from].to_i > BudgetTotal::PER_INHABITANT_FILTER_MIN || per_inhabitant_filter[:to].to_i < BudgetTotal::PER_INHABITANT_FILTER_MAX))
+          terms << {range: { amount_per_inhabitant: { gte: per_inhabitant_filter[:from].to_i, lte: per_inhabitant_filter[:to].to_i} }}
+        end
+
+        terms << {term: { autonomy_id: aarr_filter }}  unless aarr_filter.blank?
+
+        query = {
+          sort: [ { options[:variable].to_sym => { order: 'desc' } } ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms,
+                  must_not: {
+                    exists: {
+                      field: "functional_code",
+                    },
+                  },
+                  must_not: {
+                    exists: {
+                      field: "custom_code",
+                    }
+                  }
+                }
+              }
+            }
+          },
+          size: 10_000
+        }
+
+        query.merge!(size: options[:per_page]) if options[:per_page].present?
+        query.merge!(from: options[:offset]) if options[:offset].present?
+        query.merge!(_source: false) if options[:to_rank]
+
+        SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast, type: options[:area_name], body: query
+      end
+
+      def find(options)
+        return self.search(options)['hits'].detect{|h| h['code'] == options[:code] }
+      end
+
+      def compare(options)
+        terms = [{terms: { ine_code: options[:ine_codes] }},
+                 {term: { level: options[:level] }},
+                 {term: { kind: options[:kind] }},
+                 {term: { year: options[:year] }}]
+
+        terms << {term: { parent_code: options[:parent_code] }} if options[:parent_code].present?
+        terms << {term: { code: options[:code] }} if options[:code].present?
+
+        query = {
+          sort: [
+            { code: { order: 'asc' } },
+            { ine_code: { order: 'asc' }}
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          size: 10_000
+        }
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast, type: options[:type] , body: query
+        response['hits']['hits'].map{ |h| h['_source'] }
+      end
+
+      def has_children?(options)
+        options.symbolize_keys!
+        conditions = { parent_code: options[:code], level: options[:level].to_i + 1, type: options[:area] }
+        conditions.merge! options.slice(:ine_code,:kind,:year)
+
+        return search(conditions)['hits'].length > 0
+      end
+
+      def top_differences(options)
+        terms = [{term: { kind: options[:kind] }}, {term: { year: options[:year] }}, {term: { level: 3 }}]
+        terms << {term: { ine_code: options[:ine_code] }} if options[:ine_code].present?
+        terms << {term: { code: options[:code] }} if options[:code].present?
+
+        query = {
+          sort: [
+            { code: { order: 'asc' } }
+          ],
+          query: {
+            filtered: {
+              filter: {
+                bool: {
+                  must: terms
+                }
+              }
+            }
+          },
+          size: 10_000
+        }
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_forecast, type: (options[:type] || EconomicArea.area_name), body: query
+
+        planned_results = response['hits']['hits'].map{ |h| h['_source'] }
+
+        response = SearchEngine.client.search index: SearchEngineConfiguration::BudgetLine.index_executed, type: (options[:type] || EconomicArea.area_name), body: query
+
+        executed_results = response['hits']['hits'].map{ |h| h['_source'] }
+
+        results = {}
+        planned_results.each do |p|
+          if e = executed_results.detect{|e| e['code'] == p['code']}
+            results[p['code']] = [p['amount'], e['amount'], ((e['amount'].to_f - p['amount'].to_f)/p['amount'].to_f) * 100]
+          end
+        end
+
+        return results.sort{ |b, a| a[1][2] <=> b[1][2] }[0..15], results.sort{ |a, b| a[1][2] <=> b[1][2] }[0..15]
+      end
+
+      def validate_conditions(conditions)
+        if conditions.has_key?(:kind)
+          raise BudgetLine::InvalidSearchConditions unless all_kinds.include?(conditions[:kind])
+        end
+        if conditions.has_key?(:area_name)
+          raise BudgetLine::InvalidSearchConditions unless BudgetArea.all_areas_names.include?(conditions[:area_name])
+        end
+      end
+
+    end
+
+    included do
+
+      private_class_method :validate_conditions
+
+      def elastic_search_index
+        SearchEngineConfiguration::BudgetLine.send(index)
+      end
+
+      def elasticsearch_as_json
+        {
+          ine_code: ine_code,
+          province_id: province_id,
+          autonomy_id: autonomy_id,
+          year: year,
+          population: population,
+          amount: amount,
+          code: code,
+          level: level,
+          kind: kind,
+          amount_per_inhabitant: amount_per_inhabitant,
+          parent_code: parent_code
+        }
+      end
+
+    end
+
+  end
+end
